refactor: share native smoke test fixtures
Centralize identical executable-writing, smoke-repository copying, and fake Go build fixtures for six provider suites. Preserve provider-specific lifecycle and cleanup assertions, leave stricter builders separate, and exercise paths with spaces in the existing working-directory regression. Production scripts and runtime behavior are unchanged.
